COM-2016 rename style + change text in modal
